requirements.txt
kafka-python==2.0.2
requests==2.31.0
boto3==1.28.0
python-dotenv==1.0.0
pytz==2024.1
yfinance==0.2.32
pandas==2.0.0